internal: move missing unsafe diagnostic to hir
